# backend/src/main/java/com/focusflow/plan/dto/DailyPlanSummaryResponse.java
package com.focusflow.plan.dto;

import java.time.Instant;
import java.time.LocalDate;

public record DailyPlanSummaryResponse(
		Long id,
		LocalDate planDate,
		Instant createdAt,
		int itemCount,
		boolean hasWarning,
		Integer availableMinutes) {}