# docs/openapi.yaml
components:
  schemas:
    CreateTaskRequest:
      properties:
        description:
          type: "string"
        dueDate:
          format: "date"
          type: "string"
        estimatedMinutes:
          format: "int32"
          type: "integer"
        priority:
          enum:
          - "LOW"
          - "MEDIUM"
          - "HIGH"
          type: "string"
        title:
          minLength: 1
          type: "string"
      type: "object"
    DailyPlanItemResponse:
      properties:
        position:
          format: "int32"
          type: "integer"
        task:
          $ref: "#/components/schemas/TaskResponse"
      type: "object"
    DailyPlanResponse:
      properties:
        availableMinutes:
          format: "int32"
          type: "integer"
        createdAt:
          format: "date-time"
          type: "string"
        id:
          format: "int64"
          type: "integer"
        items:
          items:
            $ref: "#/components/schemas/DailyPlanItemResponse"
          type: "array"
        planDate:
          format: "date"
          type: "string"
        warning:
          $ref: "#/components/schemas/DailyPlanWarning"
      type: "object"
    DailyPlanSummaryResponse:
      properties:
        availableMinutes:
          format: "int32"
          type: "integer"
        createdAt:
          format: "date-time"
          type: "string"
        hasWarning:
          type: "boolean"
        id:
          format: "int64"
          type: "integer"
        itemCount:
          format: "int32"
          type: "integer"
        planDate:
          format: "date"
          type: "string"
      type: "object"
    DailyPlanWarning:
      properties:
        estimatedTasks:
          items:
            $ref: "#/components/schemas/EstimatedTask"
          type: "array"
        minimumAvailableMinutes:
          format: "int32"
          type: "integer"
        unestimatedTasks:
          items:
            $ref: "#/components/schemas/UnestimatedTask"
          type: "array"
      type: "object"
    EstimatedTask:
      properties:
        estimatedMinutes:
          format: "int32"
          type: "integer"
        taskId:
          format: "int64"
          type: "integer"
        title:
          type: "string"
      type: "object"
    GeneratePlanRequest:
      properties:
        availableMinutes:
          format: "int32"
          minimum: 1
          type: "integer"
        planDate:
          format: "date"
          type: "string"
      required:
      - "availableMinutes"
      - "planDate"
      type: "object"
    LoginRequest:
      properties:
        password:
          minLength: 1
          type: "string"
        username:
          minLength: 1
          type: "string"
      type: "object"
    PageResponseDailyPlanSummaryResponse:
      properties:
        content:
          items:
            $ref: "#/components/schemas/DailyPlanSummaryResponse"
          type: "array"
        page:
          format: "int32"
          type: "integer"
        size:
          format: "int32"
          type: "integer"
        totalElements:
          format: "int64"
          type: "integer"
        totalPages:
          format: "int32"
          type: "integer"
      type: "object"
    RegisterRequest:
      properties:
        email:
          minLength: 1
          type: "string"
        password:
          maxLength: 2147483647
          minLength: 8
          type: "string"
        username:
          maxLength: 30
          minLength: 3
          type: "string"
      type: "object"
    TaskResponse:
      properties:
        description:
          type: "string"
        dueDate:
          format: "date"
          type: "string"
        estimatedMinutes:
          format: "int32"
          type: "integer"
        id:
          format: "int64"
          type: "integer"
        priority:
          enum:
          - "LOW"
          - "MEDIUM"
          - "HIGH"
          type: "string"
        status:
          enum:
          - "OPEN"
          - "IN_PROGRESS"
          - "DONE"
          - "CANCELLED"
          type: "string"
        title:
          type: "string"
      type: "object"
    UnestimatedTask:
      properties:
        taskId:
          format: "int64"
          type: "integer"
        title:
          type: "string"
      type: "object"
    UpdateTaskRequest:
      properties:
        description:
          type: "string"
        dueDate:
          format: "date"
          type: "string"
        estimatedMinutes:
          format: "int32"
          type: "integer"
        priority:
          enum:
          - "LOW"
          - "MEDIUM"
          - "HIGH"
          type: "string"
        status:
          enum:
          - "OPEN"
          - "IN_PROGRESS"
          - "DONE"
          - "CANCELLED"
          type: "string"
        title:
          minLength: 1
          type: "string"
      type: "object"
    UserResponse:
      properties:
        email:
          type: "string"
        id:
          format: "int64"
          type: "integer"
        username:
          type: "string"
      type: "object"
    ValidationDto:
      properties:
        name:
          minLength: 1
          type: "string"
      type: "object"
info:
  title: "OpenAPI definition"
  version: "v0"
openapi: "3.1.0"
paths:
  /__test/errors/ai-provider:
    get:
      operationId: "aiProvider"
      responses:
        "200":
          description: "OK"
      tags:
      - "global-exception-handler-test-controller"
  /__test/errors/forbidden:
    get:
      operationId: "forbidden"
      responses:
        "200":
          description: "OK"
      tags:
      - "global-exception-handler-test-controller"
  /__test/errors/generic:
    get:
      operationId: "generic"
      responses:
        "200":
          description: "OK"
      tags:
      - "global-exception-handler-test-controller"
  /__test/errors/not-found:
    get:
      operationId: "notFound"
      responses:
        "200":
          description: "OK"
      tags:
      - "global-exception-handler-test-controller"
  /__test/errors/validation:
    post:
      operationId: "validation"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValidationDto"
        required: true
      responses:
        "200":
          description: "OK"
      tags:
      - "global-exception-handler-test-controller"
  /api/auth/login:
    post:
      operationId: "login"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
          description: "OK"
      tags:
      - "auth-controller"
  /api/auth/me:
    get:
      operationId: "me"
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
          description: "OK"
      tags:
      - "auth-controller"
  /api/auth/register:
    post:
      operationId: "register"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
        required: true
      responses:
        "201":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
          description: "Created"
      tags:
      - "auth-controller"
  /api/daily-plans:
    get:
      operationId: "list_1"
      parameters:
      - in: "query"
        name: "page"
        required: false
        schema:
          default: 0
          format: "int32"
          type: "integer"
      - in: "query"
        name: "size"
        required: false
        schema:
          default: 20
          format: "int32"
          type: "integer"
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageResponseDailyPlanSummaryResponse"
          description: "OK"
      tags:
      - "daily-plan-controller"
  /api/daily-plans/generate:
    post:
      operationId: "generate"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GeneratePlanRequest"
        required: true
      responses:
        "201":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DailyPlanResponse"
          description: "Created"
      tags:
      - "daily-plan-controller"
  /api/daily-plans/latest:
    get:
      operationId: "latest"
      parameters:
      - in: "query"
        name: "planDate"
        required: false
        schema:
          format: "date"
          type: "string"
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DailyPlanResponse"
          description: "Latest plan found"
        "204":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DailyPlanResponse"
          description: "No plan for the requested date"
      tags:
      - "daily-plan-controller"
  /api/daily-plans/{id}:
    delete:
      operationId: "delete_1"
      parameters:
      - in: "path"
        name: "id"
        required: true
        schema:
          format: "int64"
          type: "integer"
      responses:
        "204":
          description: "No Content"
      tags:
      - "daily-plan-controller"
    get:
      operationId: "getById_1"
      parameters:
      - in: "path"
        name: "id"
        required: true
        schema:
          format: "int64"
          type: "integer"
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DailyPlanResponse"
          description: "OK"
      tags:
      - "daily-plan-controller"
  /api/tasks:
    get:
      operationId: "list"
      responses:
        "200":
          content:
            '*/*':
              schema:
                items:
                  $ref: "#/components/schemas/TaskResponse"
                type: "array"
          description: "OK"
      tags:
      - "task-controller"
    post:
      operationId: "create"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
        required: true
      responses:
        "201":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TaskResponse"
          description: "Created"
      tags:
      - "task-controller"
  /api/tasks/{id}:
    delete:
      operationId: "delete"
      parameters:
      - in: "path"
        name: "id"
        required: true
        schema:
          format: "int64"
          type: "integer"
      responses:
        "204":
          description: "No Content"
      tags:
      - "task-controller"
    get:
      operationId: "getById"
      parameters:
      - in: "path"
        name: "id"
        required: true
        schema:
          format: "int64"
          type: "integer"
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TaskResponse"
          description: "OK"
      tags:
      - "task-controller"
    put:
      operationId: "update"
      parameters:
      - in: "path"
        name: "id"
        required: true
        schema:
          format: "int64"
          type: "integer"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskRequest"
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TaskResponse"
          description: "OK"
      tags:
      - "task-controller"